Fix udev rule to not start tor process twice
Details here:
https://groups.google.com/d/msgid/qubes-users/54B10566.5050005%40gmx.com
The report:
After some digging I found that, when tor is installed via the
qubes-tor-repo, for some reason the
/usr/lib/qubes-tor/start_tor_proxy.sh script is started by TWO systemd
units, at different times during the startup.

The first instance tries to start, but when the script is launched for
the second time it issues the "killall tor" at the beginning, asking
the first instance to close asap, and then launches a second instance
that cannot bind to the same ports (the first instance is still
        running, albeit closing) and exits. Then the first instance
closes
because of the "killall tor", and the TorVM remains with no tor
process active.